rename ElementNameResolver
